requirements.txt
Flask-Login
Flask-SQLAlchemy
geopy==2.*
folium==0.*
Werkzeug
gunicorn